refactor(parser-v3): Move scripts parsing in a separate ScriptParser
The main Parser subscribes to events of TemplateParser.
Events are intercepted for additional parsing. Javascript
expression found in the template are passed to the
same ScriptParser instance so that the expression can
share the data previously parsed in the scripts.
