+++ lib/v3/template.js
@@ -3,7 +3,6 @@ const { Parser: HtmlParser } = require('htmlparser2-svelte');
 
 const { parseAndMergeKeywords } = require('./v3-utils');
 const { parseComment, hasOwnProperty } = require('../utils');
-const { getAstDefaultOptions } = require('../options');
 
 const TemplateEvent = Object.freeze({
     DATA: 'data',
@@ -24,11 +23,15 @@ class TemplateParser extends EventEmitter {
         super();
 
         this.structure = parser.structure;
-        this.astOptions = getAstDefaultOptions();
         this.features = parser.features;
         this.includeSourceLocations = parser.includeSourceLocations;
 
-        this.eventsEmitted = parser.eventsEmitted;
+        // Internal properties
+        /**
+         * Map of events already emitted. Check if an event exists in this map
+         * before emitting it again.
+         */
+        this.eventsEmitted = Object.create(null); // Empty Map
     }
 
     parse() {
